requirements.txt
torch>=2.0.0
transformers>=4.57.1
accelerate>=0.20.0
huggingface-hub>=0.24.0
opencv-python>=4.10.0.0
Pillow>=10.0.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.1